gst/spectrum/gstspectrum.*: Use correct types for frame/fft counters and some minor cleanup.
Original commit message from CVS:
* gst/spectrum/gstspectrum.c: (gst_spectrum_transform_ip):
* gst/spectrum/gstspectrum.h:
Use correct types for frame/fft counters and some minor
cleanup.

+++ gst/spectrum/gstspectrum.h
@@ -49,9 +49,9 @@ struct _GstSpectrum
   guint bands;                  /* number of spectrum bands */
   gint threshold;               /* energy level treshold */
 
-  guint num_frames;             /* frame count (1 sample per channel)
+  guint64 num_frames;           /* frame count (1 sample per channel)
                                  * since last emit */
-  guint num_fft;                /* number of FFTs since last emit */
+  guint64 num_fft;              /* number of FFTs since last emit */
   GstClockTime message_ts;      /* starttime for next message */
 
   /* <private> */
